backend/pyproject.toml
[project]
name = "funke-backend"
version = "0.1.0"
description = "Funke event management backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "mangum>=0.17.0",
    "boto3>=1.34.0",
    "pydantic[email]>=2.6.0",
    "pydantic-settings>=2.2.0",
    "python-jose[cryptography]>=3.3.0",
    "httpx>=0.27.0",
    "uvicorn[standard]>=0.23.0",
]

[project.optional-dependencies]
dev = [
    "black>=23.9.0",
    "ruff>=0.1.7",
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "moto[dynamodb]>=5.0.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "B", "UP", "ASYNC", "S", "COM"]
ignore = ["S101"]

[tool.ruff.per-file-ignores]
"tests/*" = ["S101"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
